// src/app/interfaces/settings.interface.ts
export interface Settings {
  address: string;
  placeName: string;
  viewShortName: string;
  email: string;
  leaderName: string;
  leaderTitle: string;
  phone: string;
  social: object;
  logoImgUrl: string;
  featureImgs: object;
  featureEventImgs: object;
  newsletterUrl: string;
  location: {
    latitude: string,
    longitude: string
  };
  name: string;
}